+++ Makefile.am
@@ -399,7 +399,8 @@ endif
 if SUNOS
 include_HEADERS += include/uv-sunos.h
 libuv_la_CFLAGS += -D__EXTENSIONS__ -D_XOPEN_SOURCE=500
-libuv_la_SOURCES += src/unix/sunos.c
+libuv_la_SOURCES += src/unix/no-proctitle.c \
+                    src/unix/sunos.c
 endif
 
 if OS390
